feat(react-ui): rename modal folder
